split main.go and test_test.go up
